% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_heatmap.R
\name{core_heatmap}
\alias{core_heatmap}
\title{Draw the heatmap of the core members across samples}
\usage{
core_heatmap(df, abundance = 5, ubiquity = 80)
}
\arguments{
\item{df}{A data frame which represents the ASV/OTU table with raw counts. The ASVs/OTUs names should be set as the row names of the data frame.}

\item{abundance}{A number (between 0 and 100) representing the relative abundance threshold (inclusive).}

\item{ubiquity}{A number (between 0 and 100) representing the ubiquity of ASV/OTU in sample threshold (inclusive).}
}
\value{
Plot a heatmap.
}
\description{
Draw the heatmap of the core members across samples
}
\examples{
df <- data.frame(sample_x = 1:10, sample_y = 1:10)
rownames(df) <- letters[1:10]
core_heatmap(df, abundance = 10, ubiquity = 100)

}
