% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_microbiota.R
\name{core_microbiota}
\alias{core_microbiota}
\title{Get core microbiota from ASV/OTU table}
\usage{
core_microbiota(df, abundance = 5, ubiquity = 80, stats = FALSE)
}
\arguments{
\item{df}{A data frame which represents the ASV/OTU table with raw counts. The ASVs/OTUs names should be set as the row names of the data frame.}

\item{abundance}{A number (between 0 and 100) representing the relative abundance threshold (inclusive).}

\item{ubiquity}{A number (between 0 and 100) representing the ubiquity of ASV/OTU in sample threshold (inclusive).}

\item{stats}{A boolean specifying if relative abundance and ubiquity should be outputted along with ASV/OTU names.}
}
\value{
A list of ASV/OTU names belonging to the defined core microbiota.
}
\description{
Get core microbiota from ASV/OTU table
}
\examples{
df <- data.frame(sample_x = 1:10, sample_y = 1:10)
rownames(df) <- letters[1:10]
core_microbiota(df, abundance = 10, ubiquity = 100, stats = TRUE)

}
