% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_table.R
\name{core_table}
\alias{core_table}
\title{Compute relative sequence abundance and ubiquity across all samples}
\usage{
core_table(df)
}
\arguments{
\item{df}{A data frame representing a ASV/OTU table with raw counts or
phyloseq object. The ASVs/OTUs names should be set as the row
names of the data frame.}
}
\value{
A data frame with relative abundance, and ubiquity of ASVs/OTUs.
}
\description{
Compute relative sequence abundance and ubiquity across all samples
}
\examples{
df <- data.frame(sample_x = 1:10, sample_y = 1:10)
rownames(df) <- letters[1:10]
core_table(df)

}
