% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{distribution}
\alias{distribution}
\title{Draw the distribution of the relative abundance or ubiquity}
\usage{
distribution(x)
}
\arguments{
\item{x}{A data frame representing a ASV/OTU table with raw counts or
phyloseq object. The ASVs/OTUs names should be set as the row
names of the data frame.}
}
\value{
Plot a histogram of relative abundance and ubiquity.
}
\description{
Draw the distribution of the relative abundance or ubiquity
}
\examples{
df <- data.frame(sample_x = 1:10, sample_y = 1:10)
rownames(df) <- letters[1:10]
distribution(df)

}
