% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_core_table.R
\name{get_core_table}
\alias{get_core_table}
\title{Compute relative sequence abundance and ubiquity across all samples}
\usage{
get_core_table(df)
}
\arguments{
\item{df}{A data frame which represents the ASV/OTU table with raw counts. The ASVs/OTUs names should be set as the row names of the data frame.}
}
\value{
A data frame with raw counts, relative abundance, and ubiquity of ASVs/OTUs.
}
\description{
Compute relative sequence abundance and ubiquity across all samples
}
\examples{
df <- data.frame(sample_x = 1:10, sample_y = 1:10)
rownames(df) <- letters[1:10]
get_core_table(df)

}
