% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_satellite.R
\name{get_satellite}
\alias{get_satellite}
\title{Get satellite microbiota based on ASV Poisson modelling}
\usage{
get_satellite(x)
}
\arguments{
\item{x}{A data frame representing a ASV/OTU table with raw counts or
phyloseq object.}
}
\value{
A data frame of species abundance distribution of satellite species
}
\description{
Get satellite microbiota based on ASV Poisson modelling
}
\examples{
df <- data.frame(x=seq(0,10), y=seq(11,21))
get_satellite(df)

}
