% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{plot_graph}
\alias{plot_graph}
\title{Plot core/satellite graphic}
\usage{
plot_graph(x)
}
\arguments{
\item{x}{A data frame of species abundance distribution generated using
\code{sad}.}
}
\value{
A plot
}
\description{
Plot core/satellite graphic
}
\examples{
df <- data.frame(x=seq(0,10), y=seq(11,21))
sad <- sad(df)
plot_graph(sad)

}
