% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rai.R
\name{rai}
\alias{rai}
\title{Compute rarefaction aware indice}
\usage{
rai(rarefied_microbiota, non_rarefied_microbiota)
}
\arguments{
\item{rarefied_microbiota}{A vector of core microbiota members}

\item{non_rarefied_microbiota}{A vector of core microbiota members}
}
\value{
A rarefaction aware index
}
\description{
Compute rarefaction aware indice
}
