% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefaction_aware_index.R
\name{rarefaction_aware_index}
\alias{rarefaction_aware_index}
\title{Compute rarefaction aware indice}
\usage{
rarefaction_aware_index(rarefied_microbiota, non_rarefied_microbiota)
}
\arguments{
\item{rarefied_microbiota}{A vector of core microbiota members}

\item{non_rarefied_microbiota}{A vector of core microbiota members}
}
\value{
An rarefaction aware index
}
\description{
Compute rarefaction aware indice
}
