% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sad.R
\name{sad}
\alias{sad}
\title{Modelling species abundance distribution of ASV/OTU following Poisson
distribution}
\usage{
sad(x, taxa_are_rows = FALSE)
}
\arguments{
\item{x}{A data frame representing a ASV/OTU table with raw counts or
phyloseq object.}

\item{taxa_are_rows}{A single logical specifying the orientation of
the abundance table.}
}
\value{
A data frame of species abundance distribution
}
\description{
Modelling species abundance distribution of ASV/OTU following Poisson
distribution
}
\examples{
df <- data.frame(x=seq(0,10), y = seq(11,21))
sad(df)

}
