% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sniff.R
\name{sniff}
\alias{sniff}
\title{Sniff the number of members of the core microbiota given the parameters}
\usage{
sniff(x, abundance = 0.1, ubiquity = 0.8)
}
\arguments{
\item{x}{A data frame representing a ASV/OTU table with raw counts or
phyloseq object. The ASVs/OTUs names should be set as the row
names of the data frame.}

\item{abundance}{A number (between 0 and 100) or NULL representing the
relative abundance threshold (inclusive).}

\item{ubiquity}{A number (between 0 and 100) or NULL representing the
ubiquity of ASV/OTU in sample threshold (inclusive).}
}
\value{
The number of members of the core microbiota given the parameters.
}
\description{
Sniff the number of members of the core microbiota given the parameters
}
\examples{
df <- data.frame(sample_x = 1:10, sample_y = 1:10)
rownames(df) <- paste0("OTU", "_", 1:10)
sniff(df)

}
