% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stats}
\alias{stats}
\title{Compute total relative abundance and ubiquity for all samples}
\usage{
stats(x)
}
\arguments{
\item{x}{A data frame representing a ASV/OTU table with raw counts or
phyloseq object. The ASVs/OTUs names should be set as the row
names of the data frame.}
}
\value{
A data frame with raw counts, relative abundance, and ubiquity of
        ASVs/OTUs.
}
\description{
Compute total relative abundance and ubiquity for all samples
}
\examples{
df <- data.frame(sample_x = 1:10, sample_y = 1:10)
rownames(df) <- letters[1:10]
stats(df)

}
