% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_heatmap.R
\name{core_heatmap}
\alias{core_heatmap}
\title{Draw the heatmap of the core members across samples}
\usage{
core_heatmap(x, pheatmap = FALSE)
}
\arguments{
\item{x}{A data frame representing a ASV/OTU table with raw counts or
phyloseq object. The ASVs/OTUs names should be set as the row
names of the data frame.}

\item{pheatmap}{A boolean specifying the use or not of pheatmap to draw
the heatmap. Default is set to FALSE.}
}
\value{
Plot a heatmap.
}
\description{
Draw the heatmap of the core members across samples
}
\examples{
df <- data.frame(sample_x = 1:10, sample_y = 1:10)
rownames(df) <- letters[1:10]
core_heatmap(get_core_table(df))

}
