% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_microbiota.R
\name{core_microbiota}
\alias{core_microbiota}
\title{Get core microbiota from ASV/OTU table}
\usage{
core_microbiota(x, abundance = 0.1, ubiquity = 0.8, stats = FALSE)
}
\arguments{
\item{x}{A data frame representing a ASV/OTU table with raw counts or
phyloseq object. The ASVs/OTUs names should be set as the row
names of the data frame.}

\item{abundance}{A number (between 0 and 100) or NULL representing the
relative abundance threshold (inclusive).}

\item{ubiquity}{A number (between 0 and 100) or NULL representing the
ubiquity of ASV/OTU in sample threshold (inclusive).}

\item{stats}{A boolean value (default FALSE) to enable outputting of relative
abundance and ubiquity along ASVs/OTUs names.}
}
\value{
A list of ASV/OTU names belonging to the defined core microbiota.
}
\description{
Get core microbiota from ASV/OTU table
}
\examples{
df <- data.frame(sample_x = 1:10, sample_y = 1:10)
rownames(df) <- letters[1:10]
core_microbiota(df, abundance = 0.01, ubiquity = 1, stats = TRUE)

}
