% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{distribution}
\alias{distribution}
\title{Draw the distribution of the relative abundance or ubiquity}
\usage{
distribution(df, type = "abundance")
}
\arguments{
\item{df}{A data frame which represents the ASV/OTU table with raw counts. The ASVs/OTUs names should be set as the row names of the data frame.}

\item{type}{A character string specifying the type of histogram to plot. Can be either abundance or ubiquity.}
}
\value{
Plot a histogram.
}
\description{
Draw the distribution of the relative abundance or ubiquity
}
\examples{
df <- data.frame(sample_x = 1:10, sample_y = 1:10)
rownames(df) <- letters[1:10]
distribution(df, type = "abundance")

}
